requirements.txt
Django==4.2.1
djangorestframework==3.12.4
drf_yasg==2.2.0
asgiref==3.6.0
attrs==23.1.0
auth==0.5.3
autobahn==23.6.2
Automat==22.10.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
Django==4.2.1
django-cors-headers==4.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
dnspython==2.3.0
eventlet==0.33.3
falcon==3.1.1
greenlet==2.0.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
MarkupSafe==2.1.3
mongoengine==0.27.0
msgpack==1.0.5
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.4.0
pyOpenSSL==23.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
zope.interface==6.0

